style(core): added zero opacity to the table’s hidden div meant for calculating the length of the la
affects: @medly-components/core

There’s a hidden div used internally to calculate the length of the column’s largest text to provide
proper column expanding feature. This div was visible on the viewport when the text was longer. In
this fix, we added opacity to the hidden div, so that it is not visible in the viewport.